General changes and desingn decisions in rotations
Rotations algorithms around a vortex have been deleted due to bad
efficiency, they waste too much time to do a single action. Now,
rotations that are not around a single element of the Tetrimino, will be
hardcoded, I do not like this solution, but it's the best in terms of
efficiency. I will see if I can make a better algorithm in the futere.